Refactor: `MachineIdDistributor`
